move Zod schemas to separate file
